+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## X.Y.Z (unreleased)
 
+### Improvements
+
+- Added `nixpkgs.rocmSupport` option to enable ROCm support in nixpkgs configuration.
+
 ## 2.0.6 (2026-03-22)
 
 ### Bug Fixes